+++ CHANGES.md
@@ -2,6 +2,7 @@
 ======
 * Fixed an issue with Java 5 support.
 * Reopened all previously closed editors after applying a theme, not just some.
+* Made color themes apply correctly in open CDT editors.
 
 0.11.0 (2011-04-28)
 ===================